Handle empty data URIs; Match both ogg and mp3 files.
When configuring custom sounds, if some entries are left empty,
then an empty "data:" URI will be returned and there will be no
sound at all. This commit changes it so that the default Lichess
sounds will still be played.

Lichess can serve either the ogg or the mp3 audio files. Change
urlFilter to regexFilter to match both versions.